# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Exam generator API",
        "description": "API for exam generator project at UNSA",
        "contact": {
            "email": "rhualla@unsa.edu.pe"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "https://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.0.0",
        "x-logo": {
            "url": "https://via.placeholder.com/190x90.png?text=L5-Swagger"
        }
    },
    "servers": [
        {
            "url": "/var/www/html/exam_generator/backend/app"
        }
    ],
    "paths": {
...
        "/matrices": {
            "get": {
                "tags": [
                    "Matrices"
                ],
                "summary": "Listar matrices",
                "description": "Obtiene la lista completa de matrices",
                "operationId": "8be866440772b6e376711d93ef2e7148",
                "responses": {
                    "200": {
                        "description": "Lista de matrices",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/Matrix"
                                    }
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "No autenticado"
                    }
                },
                "security": [
                    {
                        "sanctum_token": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Matrices"
                ],
                "summary": "Crear nueva matriz",
                "description": "Crea una nueva matriz y la devuelve",
                "operationId": "6fd04f8f69d8effb05d6aa5296188635",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "year",
                                    "process_id",
                                    "total_alternatives"
                                ],
                                "properties": {
                                    "year": {
                                        "type": "string",
                                        "example": "2025"
                                    },
                                    "process_id": {
                                        "type": "integer",
                                        "example": 1
                                    },
                                    "total_alternatives": {
                                        "type": "integer",
                                        "example": 100
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Matriz creada exitosamente",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Matrix"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Error de validación"
                    }
                },
                "security": [
                    {
                        "sanctum_token": []
                    }
                ]
            }
        },
        "/matrices/{id}": {
            "get": {
                "tags": [
                    "Matrices"
                ],
                "summary": "Obtener una matriz",
                "description": "Devuelve los detalles de una matriz específica",
                "operationId": "ba053eda726806b3872a06bc01de6d92",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID de la matriz",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Datos de la matriz",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Matrix"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Matriz no encontrada"
                    }
                },
                "security": [
                    {
                        "sanctum_token": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "Matrices"
                ],
                "summary": "Eliminar una matriz",
                "description": "Elimina una matriz por su ID",
                "operationId": "4167bfb511a44ec6afa50356c9fc3d18",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID de la matriz",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "204": {
                        "description": "Matriz eliminada correctamente"
                    },
                    "404": {
                        "description": "Matriz no encontrada"
                    }
                },
                "security": [
                    {
                        "sanctum_token": []
                    }
                ]
            },
            "patch": {
                "tags": [
                    "Matrices"
                ],
                "summary": "Actualizar una matriz",
                "description": "Actualiza los datos de una matriz existente",
                "operationId": "04ee4da65cb5a1b1df5a2a9c9344ecc3",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID de la matriz",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": false,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "year": {
                                        "type": "string",
                                        "example": "2025"
                                    },
                                    "process_id": {
                                        "type": "integer",
                                        "example": 1
                                    },
                                    "total_alternatives": {
                                        "type": "integer",
                                        "example": 150
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Matriz actualizada exitosamente",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Matrix"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Error de validación"
                    },
                    "404": {
                        "description": "Matriz no encontrada"
                    }
                },
                "security": [
                    {
                        "sanctum_token": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "schemas": {
            "Matrix": {
                "title": "Matrix",
                "required": [
                    "id",
                    "year",
                    "process_id",
                    "total_alternatives"
                ],
                "properties": {
                    "id": {
                        "type": "integer",
                        "example": 1
                    },
                    "year": {
                        "type": "string",
                        "example": "2025"
                    },
                    "process_id": {
                        "type": "integer",
                        "example": 2
                    },
                    "total_alternatives": {
                        "type": "integer",
                        "example": 120
                    },
                    "created_at": {
                        "type": "string",
                        "format": "date-time",
                        "example": "2024-06-27T10:00:00Z"
                    },
                    "updated_at": {
                        "type": "string",
                        "format": "date-time",
                        "example": "2024-06-28T12:30:00Z"
                    },
                    "process": {
                        "description": "Proceso relacionado",
                        "properties": {
                            "id": {
                                "type": "integer",
                                "example": 2
                            },
                            "name": {
                                "type": "string",
                                "example": "Proceso 2025"
                            }
                        },
                        "type": "object"
                    },
                    "details": {
                        "description": "Detalles de la matriz",
                        "type": "array",
                        "items": {
                            "properties": {
                                "id": {
                                    "type": "integer",
                                    "example": 10
                                },
                                "matrix_id": {
                                    "type": "integer",
                                    "example": 1
                                },
                                "label": {
                                    "type": "string",
                                    "example": "Matemática"
                                },
                                "value": {
                                    "type": "number",
                                    "format": "float",
                                    "example": 5.25
                                }
                            },
                            "type": "object"
                        }
                    }
                },
                "type": "object"
            }
        }
    },
...
}